Refactor Client test logic into into reusable Components factory class and extension methods.
