+++ src/LodeRunner.API.Test/IntegrationTests/HttpClientExtension.cs
@@ -1,0 +1,161 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Net;
+using System.Net.Http;
+using System.Net.Http.Json;
+using System.Text;
+using System.Text.Json;
+using System.Threading.Tasks;
+using LodeRunner.API.Models;
+using LodeRunner.Core.Models;
+using Xunit;
+
+namespace LodeRunner.API.Test.IntegrationTests
+{
+    /// <summary>
+    /// Http Client Extensions.
+    /// </summary>
+    public static class HttpClientExtension
+    {
+        private const int WaitingTimeIncrementMs = 5000;
+
+        /// <summary>
+        /// Send GetById Request.
+        /// </summary>
+        /// <param name="httpClient">the httpClient.</param>
+        /// <param name="clientsByIdUri">the clientsById Uri.</param>
+        /// <param name="clientStatusId">the clientStatusId.</param>
+        /// <returns>HttpResponseMessage.</returns>
+        public static async Task<HttpResponseMessage> SendGetByIdRequest(this HttpClient httpClient, string clientsByIdUri, string clientStatusId)
+        {
+            HttpResponseMessage httpResponse = await httpClient.GetAsync($"{clientsByIdUri}{clientStatusId}");
+
+            Assert.False(httpResponse.StatusCode == HttpStatusCode.NoContent, "Response Code 204 - No Content.");
+
+            return httpResponse;
+        }
+
+        /// <summary>
+        /// Waits the and validate client status.
+        /// </summary>
+        /// <param name="httpClient">The HTTP client.</param>
+        /// <param name="clientsByIdUri">The clients by identifier URI.</param>
+        /// <param name="clientStatusId">The client status identifier.</param>
+        /// <param name="clientStatusType">Type of the client status.</param>
+        /// <param name="jsonOptions">The json options.</param>
+        /// <param name="timeoutLimitMs">The timeout limit ms.</param>
+        /// <returns>The task.</returns>
+        public static async Task<bool> WaitAndValidateGetByIdForStatus(this HttpClient httpClient, string clientsByIdUri, string clientStatusId, ClientStatusType clientStatusType, JsonSerializerOptions jsonOptions, int timeoutLimitMs = 10000)
+        {
+            int timeout = WaitingTimeIncrementMs;
+
+            bool foundAndValid = false;
+
+            HttpStatusCode lastHttpCode = HttpStatusCode.Unused;
+
+            while (timeout <= timeoutLimitMs)
+            {
+                await Task.Delay(WaitingTimeIncrementMs);
+
+                // Cache
+                var httpResponse = await SendGetByIdRequest(httpClient, clientsByIdUri, clientStatusId);
+
+                lastHttpCode = httpResponse.StatusCode;
+
+                if (httpResponse.StatusCode == HttpStatusCode.NotFound)
+                {
+                    await Task.Delay(WaitingTimeIncrementMs);
+                    timeout += WaitingTimeIncrementMs;
+                }
+                else if (httpResponse.IsSuccessStatusCode)
+                {
+                    var client = await httpResponse.Content.ReadFromJsonAsync<Client>(jsonOptions);
+
+                    foundAndValid = client != null && client.Status == clientStatusType;
+
+                    if (foundAndValid)
+                    {
+                        Console.WriteLine($"Timeout:{timeout}");
+                        break;
+                    }
+                    else
+                    {
+                        await Task.Delay(WaitingTimeIncrementMs);
+                        timeout += WaitingTimeIncrementMs;
+                    }
+                }
+                else
+                {
+                    // break while loop response was not successful.
+                    break;
+                }
+            }
+
+            Assert.True(foundAndValid, $"Unable to process request for ClientStatusId: {clientStatusId}\tClientStatusType: '{clientStatusType}' - HttpStatusCode:{lastHttpCode} - Timeout:{timeout}");
+
+            return foundAndValid;
+        }
+
+        /// <summary>
+        /// ValidateResponse based on at least one ClientStatus item found.
+        /// </summary>
+        /// <param name="httpClient">the httpClient.</param>
+        /// <param name="clientsUri">clientsById Uri.</param>
+        /// <param name="clientStatusId">The client status identifier.</param>
+        /// <param name="jsonOptions">The json options.</param>
+        /// <param name="timeoutLimitMs">The timeout limit ms.</param>
+        /// <returns>the task.</returns>
+        public static async Task<bool> WaitAndValidateGetClientsForId(this HttpClient httpClient, string clientsUri, string clientStatusId, JsonSerializerOptions jsonOptions, int timeoutLimitMs = 10000)
+        {
+            int timeout = WaitingTimeIncrementMs;
+
+            bool found = false;
+
+            HttpStatusCode lastHttpCode = HttpStatusCode.Unused;
+
+            while (timeout <= timeoutLimitMs)
+            {
+                await Task.Delay(WaitingTimeIncrementMs);
+
+                var httpResponse = await httpClient.GetAsync(clientsUri);
+
+                lastHttpCode = httpResponse.StatusCode;
+
+                if (httpResponse.StatusCode == HttpStatusCode.NoContent)
+                {
+                    await Task.Delay(WaitingTimeIncrementMs);
+                    timeout += WaitingTimeIncrementMs;
+                }
+                else if (httpResponse.IsSuccessStatusCode)
+                {
+                    var clients = await httpResponse.Content.ReadFromJsonAsync<IEnumerable<Client>>(jsonOptions);
+
+                    found = clients.Any(x => x.ClientStatusId == clientStatusId);
+
+                    if (found)
+                    {
+                        break;
+                    }
+                    else
+                    {
+                        await Task.Delay(WaitingTimeIncrementMs);
+                        timeout += WaitingTimeIncrementMs;
+                    }
+                }
+                else
+                {
+                    // break while loop response was not successful.
+                    break;
+                }
+            }
+
+            Assert.True(found, $"Unable to process GetClients request, it could not verify ClientStatusId: {clientStatusId}\t- HttpStatusCode:{lastHttpCode} - Timeout:{timeout}");
+
+            return found;
+        }
+    }
+}
